Adjusting header behaviour based on the page
For that, we are overwriting the `header.html` and `script.js`
files from the theme. The idea is to decide based on the
current URL if we are on the home page or a content page.
Based on that, the `script.js` decides which `css` class to
use and then the header and navigation bar will have different
colors (the Sauce logo will also change).